finished new about page/moved userdata to settings
