+++ app/lib/view/about/about.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:url_launcher/url_launcher.dart';
+import 'package:package_info_plus/package_info_plus.dart';
 
 class AboutScreen extends StatefulWidget {
   const AboutScreen({super.key});
@@ -9,16 +10,43 @@ class AboutScreen extends StatefulWidget {
 }
 
 class _AboutScreenState extends State<AboutScreen> {
+  String? appName;
+  String? packageName;
+  String? version;
+  String? buildNumber;
+
+  @override
+  void initState() {
+    PackageInfo.fromPlatform().then((PackageInfo packageInfo){
+
+      appName = packageInfo.appName;
+      packageName = packageInfo.packageName;
+      version = packageInfo.version;
+      buildNumber = packageInfo.buildNumber;
+    });
+    super.initState();
+  }
+
   @override
   Widget build(BuildContext context) {
     return Scaffold(
+      appBar: AppBar(
+        title: const Text("Über SPH-Vertretungsplan")
+      ),
       body: ListView(
         children: <Widget>[
           ListTile(
             leading: const Icon(Icons.person),
-            title: const Text('Made with love by Alessio Caputo'),
+            title: const Text('Made with love by @alessioc42'),
             onTap: () {
               launchUrl(Uri.parse("https://github.com/alessioC42"));
+            },
+          ),
+          ListTile(
+            leading: const Icon(Icons.person),
+            title: const Text('Made with love by @kurwjan'),
+            onTap: () {
+              launchUrl(Uri.parse("https://github.com/kurwjan"));
             },
           ),
           ListTile(
@@ -45,6 +73,28 @@ class _AboutScreenState extends State<AboutScreen> {
                   "https://github.com/alessioC42/SPH-vertretungsplan"));
             },
           ),
+          ListTile(
+            leading: const Icon(Icons.info),
+            title: const Text('Open Source Lizenzen'),
+            onTap: () {
+              showLicensePage(context: context);
+            }
+          ),
+          ListTile(
+            leading: const Icon(Icons.build),
+            title: const Text("Build Informationen"),
+            onTap: (){
+              showDialog(
+                  context: context,
+                  builder: (BuildContext context) {
+                    return AlertDialog(
+                      title: const Text("App Informationen"),
+                      content: Text("appName: $appName\npackageName: $packageName\nversion: $version\nbuildNumber: $buildNumber"),
+                    );
+                  }
+              );
+            },
+          )
         ],
       ),
     );
